Managed Upload: Support resuming streams
Previously we would discard the read chunks of a stream on success or
failure of a part upload. This patch will make it discard the chunk
only on success so that when `leavePartsOnError` is set to `true`, these
parts can be reuploaded with the saved chunks still in memory.

There are also changes with how parts are queued when waiting for the
`createMultipartUpload` request. They're queued using the `self.parts`
object instead of via event handlers. This makes retry possible if the
`createMultipartUpload` request fails.

This patch also fixes an inaccuracy with `totalUploadedBytes` when
uploads are aborted and resumed. Since we're retrying parts that have
failed, `totalUploadedBytes` should be reset to the last known
successfully uploaded bytes. We track that in the
`totalSuccessfullyUploadedBytes` property.

Also I am not sure why we call `stream.resume()` in the cleanup step.
This will drain the stream and will make it unreadable when the upload
is resumed.